Remove nested None values before validating
